// app/src/pages/trace/__generated__/SpanFeedbackQuery.graphql.ts
export type AnnotatorKind = "HUMAN" | "LLM";
export type SpanFeedbackQuery$variables = {
  spanId: string;
};
export type SpanFeedbackQuery$data = {
  readonly span: {
    readonly spanAnnotations?: ReadonlyArray<{
      readonly annotatorKind: AnnotatorKind;
      readonly explanation: string | null;
      readonly label: string | null;
      readonly name: string;
      readonly score: number | null;
    }>;
  };
};
export type SpanFeedbackQuery = {
  response: SpanFeedbackQuery$data;
  variables: SpanFeedbackQuery$variables;
};